Refactoring
Refactored code and enhanced documentation for better readability and maintainability

Following changes are done:

In SVM/SVM.py, docstrings were enhanced for all functions including constructor. This provides clear understanding of parameters and return types. Additionally, up to date comments and unnecessary commented-out sections in the file are removed for clarity.

In DT/DT.py, the static method prepare_data was moved outside of the decision tree (DT) class. This is done as this method doesn't depend on the state of the DT object and hence is more fitting as a standalone function.

In requirements.txt, a new dependency 'pyqt6' was added. Looks like the team is planning to implement a GUI for their application.

In test_DT.py and main.py, the call to the DT.prepare_data method was changed to a call to the standalone prepare_data method.

In main.py, Qapplication was initialized, which is necessary for the creation of QT related objects.

This refactoring and enhancement makes the code more readable, maintainable and easier to understand. These changes should have positive impact on future development and debugging efforts.