"Improve code quality and update file paths
Various code quality improvements were made including but not limited to refactoring comments to adhere to PEP 8 style guide and improving function names for better readability. The file paths in 'test_DT.py' and 'test_SVM.py' were also updated to reference the correct directories. A new function was added to 'MainWindow.py' for dealing with closing events. In 'Data/DataPrep.py' staticmethod decorator was added to the functions that don't use self and spelled out the names of those methods for improved readability. Changes in the 'Interface/MainWindow.py' were done to create a basic UI for the application. These changes help in enhancing readability and maintainability of the code."