Add new implementation for KMeans model, Rename and modify DataPrep
KMeans provides an alternative to the more complex FKM model, which is defined from scratch. As for DataPrep, additional functionality to toggle resampling and string conversion is added and some methods are renamed for clarity. All models now also use their own evaluation methods, which conduct additional metrics calculation and reporting. New files related to correlation testing were also introduced. This change was made because FKM implementation process is strenuous hence this is a sort of 'Plan B' in case FKM does not yield the desired outcome. As for DataPrep, it provides flexibility to toggle extras which could speed up or optimise the ML training. Lastly, having evaluation metrics be part of the models themselves allows direct and clear access to these metrics.